content/html/content/test/test_bug3348.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=3348
-->
<head>
  <title>Test for Bug 3348</title>
  <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>        
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=3348">Mozilla Bug 3348</a>
<p id="display"></p>
<div id="content" style="display: none">

<form id="form1">
<input type="button" value="click here" onclick="buttonClick();">
</form>
  
</div>
<pre id="test">
<script class="testbody" type="text/javascript">

/** Test for Bug 3348 **/

var oForm = document.getElementById("form1");
is(oForm.tagName, "FORM", "tagName of HTML element gives tag in upper case");

</script>
</pre>
</body>
</html>
